Split media MDE toolbar from MDE toolbar
