+++ src/OrchardCore.Modules/OrchardCore.Markdown/Assets.json
@@ -4,18 +4,24 @@
     "generateRTL": true,
     "name": "markdown-css",
     "source": "Assets/Styles/mde.scss",
-    "tags": ["admin", "markdown", "css"]
+    "tags": [ "admin", "markdown", "css" ]
   },
   {
     "action": "copy",
     "name": "markdown-easymde",
     "source": "../../../node_modules/easymde/dist/easymde.min.js",
-    "tags": ["admin", "markdown", "js"]
+    "tags": [ "admin", "markdown", "js" ]
+  },
+  {
+    "action": "min",
+    "name": "markdown-toolbar",
+    "source": "Assets/Scripts/mde.toolbar.js",
+    "tags": [ "admin", "markdown", "js" ]
   },
   {
     "action": "min",
     "name": "markdown-mediatoolbar",
     "source": "Assets/Scripts/mde.mediatoolbar.js",
-    "tags": ["admin", "markdown", "js"]
+    "tags": [ "admin", "markdown", "js" ]
   }
 ]
